Update Table/Hierarchy flow
